Simplify captureBody

+++ snapo-link-android/link-okhttp3/src/main/java/com/openai/snapo/link/okhttp3/Constants.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 
 internal val DefaultDispatcher: CoroutineDispatcher = Dispatchers.Default
-internal const val DefaultBodyPreviewBytes: Long = 4096L
-internal const val DefaultTextBodyMaxBytes: Long = 256L * 1024L
+internal const val DefaultBodyPreviewBytes: Int = 4096
+internal const val DefaultTextBodyMaxBytes: Int = 256 * 1024
 internal const val DefaultTextPreviewChars: Int = 8 * 1024
 internal const val DefaultBinaryPreviewBytes: Int = 64
